requirements.txt
Flask==2.3.3
PyMuPDF==1.23.8
python-docx==0.8.11
pandas==2.1.1
Werkzeug==2.3.7
gunicorn==21.2.0
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.0.1
requests==2.31.0
